refactor: move AnonymizeRelationshipTemplateAllocationsAllocatedByIdentityCommand specific files to correct folder
