+++ Modules/Relationships/src/Relationships.Application/RelationshipTemplates/Commands/AnonymizeRelationshipTemplateAllocationsAllocatedByIdentity/AnonymizeRelationshipTemplateAllocationsAllocatedByIdentityCommand.cs
@@ -1,7 +1,8 @@
 ﻿using Backbone.DevelopmentKit.Identity.ValueObjects;
 using MediatR;
 
-namespace Backbone.Modules.Relationships.Application.RelationshipTemplates.Commands.AnonymizeRelationshipTemplateAllocationsAllocatedByIdentityCommand;
+namespace Backbone.Modules.Relationships.Application.RelationshipTemplates.Commands.AnonymizeRelationshipTemplateAllocationsAllocatedByIdentity;
+
 public class AnonymizeRelationshipTemplateAllocationsAllocatedByIdentityCommand : IRequest
 {
     public AnonymizeRelationshipTemplateAllocationsAllocatedByIdentityCommand(IdentityAddress identityAddress)
